backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
sqlmodel==0.0.38
# PostgreSQL driver (ya incluido para conexiones de clientes)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
pydantic-settings==2.3.4
python-dotenv==1.0.1
psycopg2-binary==2.9.11
openai>=1.0.0
